Change!!!: Changed project direction
Since this project was meant to focus on learning the Notion API and serverless code, the Twitter features aren't very important. Because Twitter aren't accepting new developer applications and are going through an API change, and their older API is a pain to work with, I've changed the scope of this project to send WhatsApp messages instead.
Many changes incoming.